SCRUM-128 Init: create a seperated function for the set page title function and a helper function
