+++ Frontend/src/lib/utils/utils.ts
@@ -1,0 +1,5 @@
+export const capitalize = (str: string) => {
+  if (!str) return;
+
+  return str[0].toUpperCase() + str.slice(1);
+}
